#31495
git-svn-id: svn+ssh://10.1.90.223/SVNroot/astra/server/trunk@204293 e9a6705e-e829-0410-83f2-cebdff2db7e5

+++ src/iatci_types.cc
@@ -682,7 +682,7 @@ bool BaggageDetails::BagTagInfo::consistentWith(const BagTagInfo& bt) const
        dest()        == bt.dest() &&
        tagAccode()   == bt.tagAccode())
     {
-       return abs(tagNum() - bt.tagNum()) == 1;
+       return abs((int)tagNum() - (int)bt.tagNum()) == 1;
     }
 
     return false;
